+++ lib/loganalytics/lib/model/log-analytics-source.ts
@@ -176,6 +176,14 @@ export interface LogAnalyticsSource {
    *
    */
   "categories"?: Array<model.LogAnalyticsCategory>;
+  /**
+   * An array of REST API endpoints for log collection.
+   */
+  "endpoints"?: Array<model.LogAnalyticsEndpoint>;
+  /**
+   * A list of source properties.
+   */
+  "sourceProperties"?: Array<model.LogAnalyticsProperty>;
 }
 
 export namespace LogAnalyticsSource {
@@ -280,6 +288,16 @@ export namespace LogAnalyticsSource {
         "categories": obj.categories
           ? obj.categories.map(item => {
               return model.LogAnalyticsCategory.getJsonObj(item);
+            })
+          : undefined,
+        "endpoints": obj.endpoints
+          ? obj.endpoints.map(item => {
+              return model.LogAnalyticsEndpoint.getJsonObj(item);
+            })
+          : undefined,
+        "sourceProperties": obj.sourceProperties
+          ? obj.sourceProperties.map(item => {
+              return model.LogAnalyticsProperty.getJsonObj(item);
             })
           : undefined
       }
@@ -388,6 +406,16 @@ export namespace LogAnalyticsSource {
         "categories": obj.categories
           ? obj.categories.map(item => {
               return model.LogAnalyticsCategory.getDeserializedJsonObj(item);
+            })
+          : undefined,
+        "endpoints": obj.endpoints
+          ? obj.endpoints.map(item => {
+              return model.LogAnalyticsEndpoint.getDeserializedJsonObj(item);
+            })
+          : undefined,
+        "sourceProperties": obj.sourceProperties
+          ? obj.sourceProperties.map(item => {
+              return model.LogAnalyticsProperty.getDeserializedJsonObj(item);
             })
           : undefined
       }