+++ lib/opsi/lib/model/summarize-exadata-insight-resource-forecast-trend-aggregation.ts
@@ -44,6 +44,10 @@ export interface SummarizeExadataInsightResourceForecastTrendAggregation {
    */
   "usageUnit": SummarizeExadataInsightResourceForecastTrendAggregation.UsageUnit;
   /**
+   * Auto-ML algorithm leveraged for the forecast. Only applicable for Auto-ML forecast.
+   */
+  "selectedForecastAlgorithm"?: string;
+  /**
    * Time series patterns used in the forecasting.
    */
   "pattern": SummarizeExadataInsightResourceForecastTrendAggregation.Pattern;