+++ lib/opsi/lib/response/index.ts
@@ -26,6 +26,8 @@ import * as ChangeExadataInsightCompartmentResponse from "./change-exadata-insig
 export import ChangeExadataInsightCompartmentResponse = ChangeExadataInsightCompartmentResponse.ChangeExadataInsightCompartmentResponse;
 import * as ChangeHostInsightCompartmentResponse from "./change-host-insight-compartment-response";
 export import ChangeHostInsightCompartmentResponse = ChangeHostInsightCompartmentResponse.ChangeHostInsightCompartmentResponse;
+import * as ChangeNewsReportCompartmentResponse from "./change-news-report-compartment-response";
+export import ChangeNewsReportCompartmentResponse = ChangeNewsReportCompartmentResponse.ChangeNewsReportCompartmentResponse;
 import * as ChangeOperationsInsightsPrivateEndpointCompartmentResponse from "./change-operations-insights-private-endpoint-compartment-response";
 export import ChangeOperationsInsightsPrivateEndpointCompartmentResponse = ChangeOperationsInsightsPrivateEndpointCompartmentResponse.ChangeOperationsInsightsPrivateEndpointCompartmentResponse;
 import * as ChangeOpsiConfigurationCompartmentResponse from "./change-opsi-configuration-compartment-response";
@@ -42,6 +44,8 @@ import * as CreateExadataInsightResponse from "./create-exadata-insight-response
 export import CreateExadataInsightResponse = CreateExadataInsightResponse.CreateExadataInsightResponse;
 import * as CreateHostInsightResponse from "./create-host-insight-response";
 export import CreateHostInsightResponse = CreateHostInsightResponse.CreateHostInsightResponse;
+import * as CreateNewsReportResponse from "./create-news-report-response";
+export import CreateNewsReportResponse = CreateNewsReportResponse.CreateNewsReportResponse;
 import * as CreateOperationsInsightsPrivateEndpointResponse from "./create-operations-insights-private-endpoint-response";
 export import CreateOperationsInsightsPrivateEndpointResponse = CreateOperationsInsightsPrivateEndpointResponse.CreateOperationsInsightsPrivateEndpointResponse;
 import * as CreateOperationsInsightsWarehouseResponse from "./create-operations-insights-warehouse-response";
@@ -60,6 +64,8 @@ import * as DeleteExadataInsightResponse from "./delete-exadata-insight-response
 export import DeleteExadataInsightResponse = DeleteExadataInsightResponse.DeleteExadataInsightResponse;
 import * as DeleteHostInsightResponse from "./delete-host-insight-response";
 export import DeleteHostInsightResponse = DeleteHostInsightResponse.DeleteHostInsightResponse;
+import * as DeleteNewsReportResponse from "./delete-news-report-response";
+export import DeleteNewsReportResponse = DeleteNewsReportResponse.DeleteNewsReportResponse;
 import * as DeleteOperationsInsightsPrivateEndpointResponse from "./delete-operations-insights-private-endpoint-response";
 export import DeleteOperationsInsightsPrivateEndpointResponse = DeleteOperationsInsightsPrivateEndpointResponse.DeleteOperationsInsightsPrivateEndpointResponse;
 import * as DeleteOperationsInsightsWarehouseResponse from "./delete-operations-insights-warehouse-response";
@@ -102,6 +108,8 @@ import * as GetExadataInsightResponse from "./get-exadata-insight-response";
 export import GetExadataInsightResponse = GetExadataInsightResponse.GetExadataInsightResponse;
 import * as GetHostInsightResponse from "./get-host-insight-response";
 export import GetHostInsightResponse = GetHostInsightResponse.GetHostInsightResponse;
+import * as GetNewsReportResponse from "./get-news-report-response";
+export import GetNewsReportResponse = GetNewsReportResponse.GetNewsReportResponse;
 import * as GetOperationsInsightsPrivateEndpointResponse from "./get-operations-insights-private-endpoint-response";
 export import GetOperationsInsightsPrivateEndpointResponse = GetOperationsInsightsPrivateEndpointResponse.GetOperationsInsightsPrivateEndpointResponse;
 import * as GetOperationsInsightsWarehouseResponse from "./get-operations-insights-warehouse-response";
@@ -172,6 +180,8 @@ import * as ListImportableComputeEntitiesResponse from "./list-importable-comput
 export import ListImportableComputeEntitiesResponse = ListImportableComputeEntitiesResponse.ListImportableComputeEntitiesResponse;
 import * as ListImportableEnterpriseManagerEntitiesResponse from "./list-importable-enterprise-manager-entities-response";
 export import ListImportableEnterpriseManagerEntitiesResponse = ListImportableEnterpriseManagerEntitiesResponse.ListImportableEnterpriseManagerEntitiesResponse;
+import * as ListNewsReportsResponse from "./list-news-reports-response";
+export import ListNewsReportsResponse = ListNewsReportsResponse.ListNewsReportsResponse;
 import * as ListOperationsInsightsPrivateEndpointsResponse from "./list-operations-insights-private-endpoints-response";
 export import ListOperationsInsightsPrivateEndpointsResponse = ListOperationsInsightsPrivateEndpointsResponse.ListOperationsInsightsPrivateEndpointsResponse;
 import * as ListOperationsInsightsWarehouseUsersResponse from "./list-operations-insights-warehouse-users-response";
@@ -308,6 +318,8 @@ import * as UpdateExadataInsightResponse from "./update-exadata-insight-response
 export import UpdateExadataInsightResponse = UpdateExadataInsightResponse.UpdateExadataInsightResponse;
 import * as UpdateHostInsightResponse from "./update-host-insight-response";
 export import UpdateHostInsightResponse = UpdateHostInsightResponse.UpdateHostInsightResponse;
+import * as UpdateNewsReportResponse from "./update-news-report-response";
+export import UpdateNewsReportResponse = UpdateNewsReportResponse.UpdateNewsReportResponse;
 import * as UpdateOperationsInsightsPrivateEndpointResponse from "./update-operations-insights-private-endpoint-response";
 export import UpdateOperationsInsightsPrivateEndpointResponse = UpdateOperationsInsightsPrivateEndpointResponse.UpdateOperationsInsightsPrivateEndpointResponse;
 import * as UpdateOperationsInsightsWarehouseResponse from "./update-operations-insights-warehouse-response";